requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
pyotp==2.9.0
qrcode==7.4.2
bcrypt==4.0.1
aiosmtpd==1.4.4
python-dotenv==1.0.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
Jinja2==3.1.2
itsdangerous==2.1.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pillow==10.1.0
